refactor: 패키지 구조 변경
- 패키지 구조를 MVC패턴에 맞게, 각 역할에 맞게 패키지를 재구성
