--- src/main/java/vendingmachine/coin/CoinGenerator.java
@@ -1,7 +1,0 @@
-package vendingmachine.coin;
-
-import vendingmachine.Money;
-
-public interface CoinGenerator {
-	Coins generate(Money money);
-}
