+++ src/main/java/vendingmachine/domain/coin/CoinGenerator.java
@@ -1,0 +1,7 @@
+package vendingmachine.domain.coin;
+
+import vendingmachine.domain.money.Money;
+
+public interface CoinGenerator {
+	Coins generate(Money money);
+}
