+++ src/main/java/vendingmachine/processor/OutputProcessor.java
@@ -1,7 +1,7 @@
 package vendingmachine.processor;
 
-import vendingmachine.Money;
-import vendingmachine.coin.Coins;
+import vendingmachine.domain.money.Money;
+import vendingmachine.domain.coin.Coins;
 
 public class OutputProcessor {
 	private static final String PRINT_RETENTION_COINS = "자판기가 보유한 동전";
