+++ src/test/java/vendingmachine/domain/quantity/QuantityTest.java
@@ -1,4 +1,4 @@
-package vendingmachine.quantity;
+package vendingmachine.domain.quantity;
 
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
